requirements.txt
evdev==1.9.2
pillow==11.3.0
pyserial==3.5